requirements.txt
python-telegram-bot[rate-limiter]==20.7
aiosqlite==0.19.0
aiohttp==3.9.1
Pillow==10.1.0
# Production dependencies
sentry-sdk==1.39.0
aiohttp-basicauth==1.0.0
# Optional for high load
# psycopg2-binary==2.9.9  # PostgreSQL support
# redis==5.0.1  # Redis caching
